pyproject.toml
[project]
name = "agentify-alfworld"
version = "0.1.0"
description = "Example code of agentifying ALFWorld benchmark for standardized agent assessment."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "a2a-sdk[http-server]>=0.3.8",
    "python-dotenv>=1.0.0",
    "alfworld",
    "typer>=0.9.0",
    "uvicorn>=0.27.0",
    "litellm>=1.0.0",
    "httpx>=0.27.0",
    "textworld>=1.5.0",
    "pyyaml>=6.0",
]

[project.scripts]
agentify-alfworld = "main:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
